parsing magic ready and cmd redirections cleared

+++ srcs/tcaps_positions.c
@@ -6,7 +6,7 @@
 /*   By: vlistrat <vlistrat@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/31 09:54:24 by vlistrat          #+#    #+#             */
-/*   Updated: 2017/01/31 14:32:45 by vlistrat         ###   ########.fr       */
+/*   Updated: 2017/02/06 12:15:17 by vlistrat         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	tcaps_recalc_pos(t_env *e)
 {
 	int		tmp;
 
-	tmp = TCAPS.nb_move + ft_strlen(e->prompt);
+	tmp = TCAPS.nb_move + (int)ft_strlen(e->prompt);
 	ioctl(0, TIOCGWINSZ, &TCAPS.ws);
 	TCAPS.nb_line = (tmp / (TCAPS.ws.ws_col)) + 1;
 	if (TCAPS.nb_line)
